requirements.txt
django
markdown
django-filter
djangorestframework
psycopg2